refactor code  for warning properties and getnugetlogcodes
